+++ src/storage/store/async.rs
@@ -1,0 +1,6 @@
+pub mod filesystem_store;
+pub mod memory_store;
+pub mod object_store;
+
+#[cfg(feature = "http")]
+pub mod http_store;